Split generic-seeming functionality into their own files
